fix(web): settle shortcuts now match the active thread state
